+++ src/KnitClient.lua
@@ -42,6 +42,23 @@ type Service = {
 	[any]: any,
 }
 
+--[=[
+	@interface KnitOptions
+	.ServicePromises boolean -- Defaults to `true`
+	@within KnitClient
+
+	`ServicePromises` defaults to `true` and indicates if service methods use promises.
+]=]
+type KnitOptions = {
+	ServicePromises: boolean,
+}
+
+local defaultOptions: KnitOptions = {
+	ServicePromises = true;
+}
+
+local selectedOptions = nil
+
 
 --[=[
 	@class KnitClient
@@ -52,19 +69,19 @@ local KnitClient = {}
 --[=[
 	@prop Player Player
 	@within KnitClient
-	Reference to the LocalPlayer
+	@readonly
+	Reference to the LocalPlayer.
 ]=]
 KnitClient.Player = game:GetService("Players").LocalPlayer
 
 --[=[
-	@prop Controllers {[string]: Controller}
-	@within KnitClient
-]=]
-KnitClient.Controllers = {} :: {[string]: Controller}
-
---[=[
 	@prop Util Folder
 	@within KnitClient
+	@readonly
+	References the Util folder. Should only be accessed when using Knit as
+	a standalone module. If using Knit from Wally, modules should just be
+	pulled in via Wally instead of relying on Knit's Util folder, as this
+	folder only contains what is necessary for Knit to run in Wally mode.
 ]=]
 KnitClient.Util = script.Parent.Parent
 
@@ -72,6 +89,7 @@ local Promise = require(KnitClient.Util.Promise)
 local Comm = require(KnitClient.Util.Comm)
 local ClientComm = Comm.ClientComm
 
+local controllers: {[string]: Controller} = {}
 local services: {[string]: Service} = {}
 local servicesFolder = nil
 
@@ -81,14 +99,14 @@ local onStartedComplete = Instance.new("BindableEvent")
 
 
 local function BuildService(serviceName: string, folder: Instance): Service
-	local service = ClientComm.new(folder, true):BuildObject()
+	local service = ClientComm.new(folder, selectedOptions.ServicePromises):BuildObject()
 	services[serviceName] = service
 	return service
 end
 
 
 local function DoesControllerExist(controllerName: string): boolean
-	local controller: Controller? = KnitClient.Controllers[controllerName]
+	local controller: Controller? = controllers[controllerName]
 	return controller ~= nil
 end
 
@@ -112,7 +130,7 @@ function KnitClient.CreateController(controllerDef: ControllerDef): Controller
 	assert(#controllerDef.Name > 0, "Controller.Name must be a non-empty string")
 	assert(not DoesControllerExist(controllerDef.Name), "Controller \"" .. controllerDef.Name .. "\" already exists")
 	local controller = controllerDef :: Controller
-	KnitClient.Controllers[controller.Name] = controller
+	controllers[controller.Name] = controller
 	return controller
 end
 
@@ -127,12 +145,12 @@ end
 	```
 ]=]
 function KnitClient.AddControllers(parent: Instance): {Controller}
-	local controllers = {}
+	local addedControllers = {}
 	for _,v in ipairs(parent:GetChildren()) do
 		if not v:IsA("ModuleScript") then continue end
-		table.insert(controllers, require(v))
+		table.insert(addedControllers, require(v))
 	end
-	return controllers
+	return addedControllers
 end
 
 
@@ -142,12 +160,12 @@ end
 	Requires all the modules that are descendants of the given parent.
 ]=]
 function KnitClient.AddControllersDeep(parent: Instance): {any}
-	local controllers = {}
+	local addedControllers = {}
 	for _,v in ipairs(parent:GetDescendants()) do
 		if not v:IsA("ModuleScript") then continue end
-		table.insert(controllers, require(v))
+		table.insert(addedControllers, require(v))
 	end
-	return controllers
+	return addedControllers
 end
 
 
@@ -176,15 +194,19 @@ end
 --[=[
 	@param controllerName string
 	@return Controller?
-	Gets the controller by name. Returns `nil` if not found. This is just
-	an alias for `KnitControllers.Controllers[controllerName]`.
+	Gets the controller by name. Throws an error if the controller
+	is not found.
 ]=]
-function KnitClient.GetController(controllerName: string): Controller?
-	return KnitClient.Controllers[controllerName]
+function KnitClient.GetController(controllerName: string): Controller
+	assert(type(controllerName) == "string", "ControllerName must be a string; got " .. type(controllerName))
+	local controller = controllers[controllerName]
+	assert(controller ~= nil, " Could not find controller \"" .. controllerName .. "\". Check to verify a controller with this name exists.")
+	return controller
 end
 
 
 --[=[
+	@param options KnitOptions?
 	@return Promise
 	Starts Knit. Should only be called once per client.
 	```lua
@@ -192,8 +214,16 @@ end
 		print("Knit started!")
 	end):catch(warn)
 	```
+
+	By default, service methods exposed to the client will return promises.
+	To change this behavior, set the `ServicePromises` option to `false`:
+	```lua
+	Knit.Start({ServicePromises = false}):andThen(function()
+		print("Knit started!")
+	end):catch(warn)
+	```
 ]=]
-function KnitClient.Start()
+function KnitClient.Start(options: KnitOptions?)
 
 	if started then
 		return Promise.reject("Knit already started")
@@ -201,7 +231,12 @@ function KnitClient.Start()
 
 	started = true
 
-	local controllers = KnitClient.Controllers
+	if options == nil then
+		selectedOptions = defaultOptions
+	else
+		assert(typeof(options) == "table", "KnitOptions should be a table or nil; got " .. typeof(options))
+		selectedOptions = options
+	end
 
 	return Promise.new(function(resolve)
 
@@ -246,7 +281,7 @@ end
 	that called `Start`.
 	```lua
 	Knit.OnStart():andThen(function()
-		local MyController = Knit.Controllers.MyController
+		local MyController = Knit.GetController("MyController")
 		MyController:DoSomething()
 	end):catch(warn)
 	```